+++ js/utils/_utils.ts
@@ -1,0 +1,52 @@
+import { LitElement } from "lit";
+
+////////////////////////////////////////////////
+// Lit helpers
+////////////////////////////////////////////////
+
+function createElement(
+  tag_name: string,
+  attrs: { [key: string]: string | null }
+): HTMLElement {
+  const el = document.createElement(tag_name);
+  for (const [key, value] of Object.entries(attrs)) {
+    if (value !== null) el.setAttribute(key, value);
+  }
+  return el;
+}
+
+function createSVGIcon(icon: string): HTMLElement {
+  const parser = new DOMParser();
+  const svgDoc = parser.parseFromString(icon, "image/svg+xml");
+  return svgDoc.documentElement;
+}
+
+// https://lit.dev/docs/components/shadow-dom/#implementing-createrenderroot
+class LightElement extends LitElement {
+  createRenderRoot() {
+    return this;
+  }
+}
+////////////////////////////////////////////////
+// Shiny helpers
+////////////////////////////////////////////////
+
+export type ShinyClientMessage = {
+  message: string;
+  headline?: string;
+  status?: "error" | "info" | "warning";
+};
+
+function showShinyClientMessage({
+  headline = "",
+  message,
+  status = "warning",
+}: ShinyClientMessage): void {
+  document.dispatchEvent(
+    new CustomEvent("shiny:client-message", {
+      detail: { headline: headline, message: message, status: status },
+    })
+  );
+}
+
+export { LightElement, createElement, createSVGIcon, showShinyClientMessage, };